Trainee UI/UX Designer
I specialize in UI/UX for web and mobile applications. I turn ideas into reality to solve problems and simplify daily goals of a target audience. Also I pay attention not only to the beauty of the site, but also to functionality and details. I do full cycle design: analyse a task, prototype an interface, create visual style and implement it all into a usable and simple end product. I want to improve my skills to be more knowledgeable.